Share one DeleteFileModal for all files

+++ teccloud-server/components/DeleteFileModal.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { Dispatch, SetStateAction } from 'react';
 import Box from '@mui/material/Box';
 import Fade from '@mui/material/Fade';
 import Modal from '@mui/material/Modal';
@@ -7,18 +7,23 @@ import Stack from '@mui/material/Stack';
 import Button from '@mui/material/Button';
 import { apiServer } from '../config';
 
-interface DeleteFileModalProps {
+export interface DeletingFile {
   fileName: string;
   originalName: string;
 }
 
-const DeleteFileModal = ({ fileName, originalName }: DeleteFileModalProps) => {
-  const [open, setOpen] = useState<boolean>(false);
-  const handleClose = () => setOpen(false);
+interface DeleteFileModalProps {
+  file: DeletingFile | null;
+  setFile: Dispatch<SetStateAction<DeletingFile | null>>;
+}
+
+const DeleteFileModal = ({ file, setFile }: DeleteFileModalProps) => {
+  const handleClose = () => setFile(null);
 
   const deleteFile = async () => {
+    if (!file) return;
     try {
-      await fetch(`${apiServer}/files/${fileName}`, {
+      await fetch(`${apiServer}/files/${file.fileName}`, {
         method: 'delete',
         credentials: 'include',
       });
@@ -29,56 +34,52 @@ const DeleteFileModal = ({ fileName, originalName }: DeleteFileModalProps) => {
   };
 
   return (
-    <>
-      <Button variant='contained' color='error' onClick={() => setOpen(true)}>
-        Delete
-      </Button>
-      <Modal open={open} onClose={handleClose}>
-        <Fade in={open}>
-          <Box
-            sx={{
-              position: 'absolute',
-              top: '50%',
-              left: '50%',
-              transform: 'translate(-50%, -50%)',
-              width: 400,
-              bgcolor: 'background.paper',
-              border: '5px solid red',
-              boxShadow: 24,
-              p: 4,
-              display: 'flex',
-              flexDirection: 'column',
-              alignItems: 'center',
-            }}
+    <Modal open={file !== null} onClose={handleClose}>
+      <Fade in={file !== null}>
+        <Box
+          sx={{
+            position: 'absolute',
+            top: '50%',
+            left: '50%',
+            transform: 'translate(-50%, -50%)',
+            bgcolor: 'background.paper',
+            border: '5px solid red',
+            boxShadow: 24,
+            p: 4,
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+          }}
+        >
+          <Typography variant='h5' textAlign='center' fontWeight='bold'>
+            Are you sure you want to permanently delete this file?
+          </Typography>
+          <Typography
+            variant='h5'
+            textAlign='center'
+            fontStyle='italic'
+            sx={{ my: 4 }}
+            noWrap
+            maxWidth='500px'
           >
-            <Typography variant='h5' textAlign='center' fontWeight='bold'>
-              Are you sure you want to permanently delete this file?
-            </Typography>
-            <Typography
-              variant='h5'
-              textAlign='center'
-              fontStyle='italic'
-              sx={{ my: 4 }}
-            >
-              {originalName}
-            </Typography>
-            <Stack
-              direction='row'
-              justifyContent='flex-start'
-              alignItems='center'
-              spacing={1}
-            >
-              <Button variant='contained' color='error' onClick={deleteFile}>
-                Yes, delete now
-              </Button>
-              <Button variant='contained' color='primary' onClick={handleClose}>
-                No, go back
-              </Button>
-            </Stack>
-          </Box>
-        </Fade>
-      </Modal>
-    </>
+            {file?.originalName || '.'}
+          </Typography>
+          <Stack
+            direction='row'
+            justifyContent='flex-start'
+            alignItems='center'
+            spacing={1}
+          >
+            <Button variant='contained' color='error' onClick={deleteFile}>
+              Yes, delete now
+            </Button>
+            <Button variant='contained' color='primary' onClick={handleClose}>
+              No, go back
+            </Button>
+          </Stack>
+        </Box>
+      </Fade>
+    </Modal>
   );
 };
